Modify get_variable_r32_1d and all type components that use it
The key is the interface for this subroutine in c_FileIO_netcdf: I
changed the `var_data` parameter to have the pointer attr instead of
the alllocatable attr. This forced all type components (there are
many) that use this sub to read from a file to also have the pointer
attr. PRMS compiles.